Restructure classes into own modules
