+++ crates/egui/src/widgets/checkbox.rs
@@ -2,8 +2,10 @@ use emath::Rect;
 
 use crate::{
     Atom, AtomLayout, Atoms, Id, IntoAtoms, NumExt as _, Response, Sense, Shape, Ui, Vec2, Widget,
-    WidgetInfo, WidgetType, epaint, pos2,
-    widget_style::{CheckboxStyle, Classes, HasClasses},
+    WidgetInfo, WidgetType,
+    class::{Classes, HasClasses},
+    epaint, pos2,
+    widget_style::CheckboxStyle,
 };
 
 // TODO(emilk): allow checkbox without a text label
