+++ tests/egui_tests/tests/test_atoms.rs
@@ -133,7 +133,7 @@ fn test_atom_layout_nesting_and_direction() {
 
         let button_frame = ui
             .get_widget_style::<ButtonStyle>(&StyleArgs {
-                classes: &egui::widget_style::Classes::default(),
+                classes: &egui::class::Classes::default(),
                 state: egui::widget_style::WidgetState::Inactive,
                 ctx: ui,
                 stack: ui.stack(),
